+++ jwst/pipeline/tests/test_calwebspec2.py
@@ -3,14 +3,11 @@ import pytest
 from ..calwebb_spec2 import Spec2Pipeline
 
 
-@pytest.fixture(scope='module')
-def fake_pipeline():
-    return Spec2Pipeline()
+def test_filenotfounderror_raised(capsys):
+    # Verify the failure is in the traceback message
+    with pytest.raises(RuntimeError, match="FileNotFoundError"):
+        Spec2Pipeline().run('file_does_not_exist.fits')
 
-
-def test_filenotfounderror_raised(fake_pipeline, capsys):
-    with pytest.raises(RuntimeError):
-        fake_pipeline.run('file_does_not_extis.fits')
-
+    # Verify the failure is printed to stderr
     captured = capsys.readouterr()
     assert 'FileNotFoundError' in captured.err